adaptation to DB new structure + /register mapping

+++ src/main/java/lcu/backend/server/persistence/AidAnswerRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface AidAnswerRepository extends JpaRepository<AidAnswer, Long> {
+public interface AidAnswerRepository extends JpaRepository<AidAnswer, Integer> {
     Optional<AidAnswer> findByRequest_Id(Integer requestId);
     boolean existsByRequest_Id(Integer requestId);
 }
